Retarget for RFC8785
RFC 8785 is a lot more fleshed out. It tries to
support floating point for example.

There's also reference tests, so we know we're
implementing something at least with a known baseline.

Note that I preserved all of the existing tests
from the olpc-cjson code, because the nice thing
is that I think as long as one sticks to ASCII
keys, RFC8785 is a superset of olpc-cjson.

The sorting though is quite messy/complicated to do
as a serde serializer, because we need to:

- Store keys as Vec<u16> (and order them via BTreeMap that way)
- Carefully intercept when we have writes to strings
  for keys and store those literally
- When we go to write an object, serialize the keys as
  proper JSON strings - which means we need a recursive
  Serializer

Signed-off-by: Colin Walters <walters@verbum.org>